+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.3.0 - The Major Upgrade Update
 
 - Change the swappiness from 10% down to 1%.
+- Change the SteamOS repositories to use the new release URLs introduced with SteamOS 3.4.
 
 ## 3.2.0 - The Minimal Image Update